requirements.txt
discord
python-dotenv
google-generativeai